#412 cr:package reorganization
